Retoques finales

+++ src/main/java/ar/com/repo/APIrest/DTO/DtoLoan.java
@@ -1,40 +1,37 @@
 package ar.com.repo.APIrest.DTO;
 
-
-
-
 import java.io.Serializable;
 
 public class DtoLoan implements Serializable {
-    private long idLoan;
-    private int total;
-    private long userId;
+  private int idLoan;
+  private int total;
+  private Integer userId;
 
 
 
-    public long getId() {
-        return idLoan;
-    }
+  public int getId() {
+    return idLoan;
+  }
 
-    public void setId(long id) {
-        this.idLoan = id;
-    }
+  public void setId(int id) {
+    this.idLoan = id;
+  }
 
-    public int getTotal() {
-        return total;
-    }
+  public int getTotal() {
+    return total;
+  }
 
-    public void setTotal(int total) {
-        this.total = total;
-    }
+  public void setTotal(int total) {
+    this.total = total;
+  }
 
-    public long getUserId() {
-        return userId;
-    }
+  public Integer getUserId() {
+    return userId;
+  }
 
-    public void setUserId(long userId) {
-        this.userId = userId;
-    }
+  public void setUserId(Integer userId) {
+    this.userId = userId;
+  }
 
 
 }
