+++ src/main/java/ar/com/repo/APIrest/repository/UserRepository.java
@@ -4,8 +4,8 @@ import ar.com.repo.APIrest.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-  User findAllById(long userId);
+  User findAllById(Integer userId);
 
 }
